refactor(scrollview): move scrollview to its own directory
